+++ tests/testthat/test_input_checks_node.r
@@ -1,6 +1,6 @@
 
 test_that("name root node", {
-  expect_error(node(c("A", "B"), type="rbernoulli"))
+  expect_error(node(c(1, 2), type="rbernoulli"))
 })
 
 test_that("name root node", {
@@ -8,8 +8,8 @@ test_that("name root node", {
 })
 
 test_that("name child node", {
-  expect_error(node(c("A", "B"), type="binomial", parents="C",
-                    betas=1))
+  expect_error(node(c(1, 2), type="binomial", parents="C",
+                    betas=1, intercept=2.1))
 })
 
 test_that("type child node", {